remove the csv and jsonl analysis files (for now)
The csv and jsonl files in this format don't provide a clear benefit; to read
them, one must write a program parsing json (e.g. using the python `json`
library), which isn't any easier than writing a program to parse the conllu
(e.g. using `pyconll`).

Maybe after some discussion we will be able to come up with a more useful csv
or jsonl format. One possibility is to use the CQP format, which contains the
morpheme-to-token collapsing, though it's not clear this is *more* readable
than conllu.